requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
bson==0.5.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.5
requests==2.32.3
requests-file==2.1.0
six==1.17.0
soupsieve==2.6
tldextract==5.1.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
